Add Faraday encoding middleware
As noted in issue #14, we have found cases where the response from
Faraday is encoded as ASCII-8BIT when it should be UTF-8. It turns out
that Net::HTTP is not reliable at setting the encoding from the return
headers. Adding a Faraday middleware explicitly for this purpose
addresses the issue.